tests/wpt/tests/shared-storage/shared-storage-writable-insecure-context.tentative.http.sub.html

<!doctype html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script>
    'use strict';

    promise_test(async t => {
      try {
        new Request('./resources/simple-module.js',
                    {sharedStorageWritable: true});
      } catch (e) {
        assert_equals(e.name, 'TypeError');
        assert_equals(e.message, `Failed to construct 'Request': `
                      + `sharedStorageWritable: sharedStorage operations `
                      + `are only available in secure contexts.`);
        return;
      }
      assert_unreached("did not reject");
    }, 'sharedStorageWritable not allowed in insecure context.');
  </script>
</body>
